01-Machine-to-Machine/requirements.txt
flask
python-dotenv
auth0-python
flask-jwt-simple[asymmetric_crypto]
